LayoutTests/fast/speechsynthesis/mac/samantha-voice-available-expected.txt

This tests that we can get Samantha.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foundSamantha is true
PASS successfullyParsed is true

TEST COMPLETE
